requirements.txt
# Core
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# ML
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
openvino>=2024.0  # Optional: 1.4x speedup on Intel CPUs

# CLI
click>=8.1.0
tqdm>=4.65.0
colorama>=0.4.6

# Dev
pytest>=7.4.0
black>=23.0.0
